Removed all  material and expert types to types.ts file
